Added email variable for Income Endpoints

+++ server/src/routes/income.py
@@ -1,12 +1,14 @@
 from typing import List
 
 import pymongo
-from fastapi import APIRouter, Body, HTTPException, status
+from fastapi import APIRouter, Body, HTTPException, status, Depends
 from fastapi.responses import JSONResponse
 from fastapi.encoders import jsonable_encoder
 from pydantic.error_wrappers import ValidationError
 from datetime import date
 
+from ..middleware import get_current_active_user
+from ..models.user import User
 from .net_worth import update_net_worth
 from ..models.income import Income, UpdateIncomeModel, AddIncome
 from ..database.database import income_collection
@@ -19,9 +21,9 @@ router = APIRouter()
 @router.get(
     "/incomes/", response_description="Get all expenses", response_model=List[Income]
 )
-def get_incomes():
+def get_incomes(current_user: User = Depends(get_current_active_user)):
     if (
-        all_incomes := income_collection.find().sort(
+        all_incomes := income_collection.find({"email": current_user["email"]}).sort(
             [("date", pymongo.DESCENDING), ("updated_at", pymongo.DESCENDING)]
         )
     ).count():
@@ -33,15 +35,15 @@ def get_incomes():
 @router.get(
     "/income/{id}", response_description="Get income by id", response_model=Income
 )
-def get_income_by_id(id):
-    if (income := income_collection.find_one({"_id": id})) is not None:
+def get_income_by_id(id, current_user: User = Depends(get_current_active_user)):
+    if (income := income_collection.find_one({"_id": id, "email": current_user["email"]})) is not None:
         return income
 
     raise HTTPException(status_code=404, detail=f"Income with id {id} not found")
 
 
 @router.post("/income/", response_description="Add new income", response_model=Income)
-def create_income(income: AddIncome = Body(...)):
+def create_income(income: AddIncome = Body(...), current_user: User = Depends(get_current_active_user)):
     if (net_worth_to_update := net_worth_collection.find_one()) is None:
         raise HTTPException(status_code=404, detail=f"Net worth not found")
 
@@ -53,6 +55,7 @@ def create_income(income: AddIncome = Body(...)):
         income.exchange_rate = get_exchange_rate_to_cad(income.currency)
 
     income_dict = {k: v for k, v in income.dict().items()}
+    income_dict["email"] = current_user["email"]
 
     try:
         insert_income = Income(**income_dict)
@@ -72,7 +75,7 @@ def create_income(income: AddIncome = Body(...)):
     response_description="Update income selected by id",
     response_model=Income,
 )
-def update_income(id, income: UpdateIncomeModel = Body(...)):
+def update_income(id, income: UpdateIncomeModel = Body(...), current_user: User = Depends(get_current_active_user)):
     if (income_to_update := income_collection.find_one({"_id": id})) is None:
         raise HTTPException(status_code=404, detail=f"Expense with id {id} not found")
     if (net_worth_to_update := net_worth_collection.find_one()) is None:
@@ -88,16 +91,17 @@ def update_income(id, income: UpdateIncomeModel = Body(...)):
             income.exchange_rate = get_exchange_rate_to_cad(income.currency)
 
     income = {k: v for k, v in income.dict().items() if v is not None}
+    income["email"] = current_user["email"]
     income = jsonable_encoder(income)
 
     if len(income) >= 1:
-        update_result = income_collection.update_one({"_id": id}, {"$set": income})
+        update_result = income_collection.update_one({"_id": id, "email": current_user["email"]}, {"$set": income})
 
         if update_result.modified_count == 1:
-            if (updated_income := income_collection.find_one({"_id": id})) is not None:
+            if (updated_income := income_collection.find_one({"_id": id, "email": current_user["email"]})) is not None:
                 return updated_income
 
-    if (existing_income := income_collection.find_one({"_id": id})) is not None:
+    if (existing_income := income_collection.find_one({"_id": id, "email": current_user["email"]})) is not None:
         return existing_income
 
     raise HTTPException(status_code=404, detail=f"Income with id {id} not found")
@@ -106,11 +110,11 @@ def update_income(id, income: UpdateIncomeModel = Body(...)):
 @router.delete(
     "/income/{id}", response_description="Delete income by id", response_model=Income
 )
-def delete_income_by_id(id):
-    income_to_delete: Income = income_collection.find_one({"_id": id})
+def delete_income_by_id(id, current_user: User = Depends(get_current_active_user)):
+    income_to_delete: Income = income_collection.find_one({"_id": id, "email": current_user["email"]})
     if income_to_delete is None:
         raise HTTPException(status_code=404, detail=f"Expense with id {id} not found")
-    if (net_worth_to_update := net_worth_collection.find_one()) is None:
+    if (net_worth_to_update := net_worth_collection.find_one({"email": current_user["email"]})) is None:
         raise HTTPException(status_code=404, detail=f"Net worth not found")
 
     update_net_worth(
@@ -119,7 +123,7 @@ def delete_income_by_id(id):
         income_to_delete.date,
     )
 
-    delete_result = income_collection.delete_one({"_id": id})
+    delete_result = income_collection.delete_one({"_id": id, "email": current_user["email"]})
 
     if delete_result.deleted_count == 1:
         return JSONResponse(
@@ -135,9 +139,9 @@ def delete_income_by_id(id):
     response_description="Returns limited number of incomes sorted by date",
     response_model=List[Income],
 )
-def limited_income(limit: int = 10, offset: int = 0):
+def limited_income(limit: int = 10, offset: int = 0, current_user: User = Depends(get_current_active_user)):
     if (
-        all_incomes := income_collection.find(limit=limit, skip=offset).sort(
+        all_incomes := income_collection.find({"email": current_user["email"]}, limit=limit, skip=offset).sort(
             [("date", pymongo.DESCENDING), ("updated_at", pymongo.DESCENDING)]
         )
     ).count(with_limit_and_skip=True):
@@ -157,10 +161,10 @@ def limited_income(limit: int = 10, offset: int = 0):
     response_description="Returns incomes that have a date between the start date and end date",
     response_model=List[Income],
 )
-def ranged_income(start_time: date, end_time: date):
+def ranged_income(start_time: date, end_time: date, current_user: User = Depends(get_current_active_user)):
     if (
         incomes := income_collection.find(
-            {"date": {"$gte": str(start_time), "$lt": str(end_time)}}
+            {"date": {"$gte": str(start_time), "$lt": str(end_time)}, "email": current_user["email"]}
         ).sort([("date", pymongo.DESCENDING), ("updated_at", pymongo.DESCENDING)])
     ).count():
         return [jsonable_encoder(next(incomes)) for _ in range(incomes.count())]
@@ -184,6 +188,7 @@ def specified_incomes(
     income_currency: str = None,
     income_exchange_rate: float = None,
     income_category: str = None,
+    current_user: User = Depends(get_current_active_user)
 ):
 
     specified_income = {}
@@ -201,6 +206,7 @@ def specified_incomes(
         specified_income["exchange_rate"] = income_exchange_rate
     if income_category:
         specified_income["currency"] = income_category
+    specified_income["email"] = current_user["email"]
 
     if (
         incomes := income_collection.find(specified_income).sort(
